split itsystemv2 tests into three responsibility areas
