+++ Tests.Integration.Presentation.Web/Tests.Integration.Presentation.Web.csproj
@@ -133,6 +133,9 @@
     <Compile Include="Contract\ItContractTest.cs" />
     <Compile Include="Deltas\V2\DeltaFeedApiV2Test.cs" />
     <Compile Include="Internal\Messages\PublicMessagesApiV2Test.cs" />
+    <Compile Include="ItSystem\V2\BaseItSystemsApiV2Test.cs" />
+    <Compile Include="ItSystem\V2\ItSystemsInternalApiV2Test.cs" />
+    <Compile Include="ItSystem\V2\ItSystemsApiV2RightsHoldersTest.cs" />
     <Compile Include="Options\V1\OptionApiTests.cs" />
     <Compile Include="Users\OrganizationRightApiTest.cs" />
     <Compile Include="Organizations\OrganizationUnitTests.cs" />
